plugins/slack/pyproject.toml
[project]
name = "gestalt-slack"
version = "0.0.1a1"
requires-python = "==3.14.*"
dependencies = ["gestalt-sdk==0.0.1a11"]

[dependency-groups]
dev = [
  "ruff==0.15.9",
  "ty==0.0.29",
  "vulture==2.16",
]

[tool.uv]
package = false

[tool.gestalt]
plugin = "provider"

[tool.gestalt.ci]
vulture-ignore-names = [
  "post_connect",
  "conversations_get_message",
  "conversations_find_user_mentions",
  "conversations_get_thread_participants",
]
